QSDK:CP Add max_oper_chwidth 0 for bandwidth EHT20 and EHT40
disabled_vht is set to 1 for the 2G mesh in both EHT20 and EHT40
bandwidths. This configuration is not populating the VHT capabilities
in the association, which is negatively impacting the KPI in the 2G mesh
instead need to add max_oper_chwidth 0 for EHT20 and EHT40 bandwidths.

Patch-dependency: None
Patch-work: None

Change-Id: Ibf21d79f9c0658d8a0f1298a8ded8da3076f1157
Signed-off-by: Harish Rachakonda <quic_rachakon@quicinc.com>
(cherry picked from commit 250f4f55dd47aaf91909e372c85d0375f4ff262b)