QSDK:CP:mac80211:Fix monitor cleanup during failure
During monitor vdev creation, when assigning channel context fails for
one hardware, expectation is deletion of monitor interface created for
all the hardware. But in failure case, the monitor_sdata is initialized
to NULL and then goes for a NULL check which will fail causing the iterface
not to be removed through drv_remove_interface(). Also
monitor_sdata_created is set to false making it skip the deletion even in
ieee80211_del_virtual_monitor().

Handle the failure case, by removing the interface properly. While at
it, also sanitize the other failure cases. Because the failure case will
not release the channel assigned for the monitor interfaces created for
other hardware in the wiphy.

Patch-dependency: 723-mac80211-add-mlo-monitor-support.patch

Note: The patch is QSDK because the dependent patch will not be
upstreamed since MLO monitor is already upstreamed in different design
by another author.

Change-Id: Ibff88c328b02cfa16fbfe625ef2ca16e155df28d
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>