QSDK:PROTOCOL: wifi: mac80211: always process 240 MHz IE
Currently 240 MHz vendor IE is processed only when VHT/HE
bandwidths are 160 MHz but in some cases these can be <= 80 MHz.
For example, if the primary channel is between 132 and 144,
bandwidth configured for legacy modes is only 80 MHz as channels
from 145 to 148 are disabled.
In such a scenario, client is connecting only on 80 MHz because
240 MHz IE is not processed. Remove the check for 160 MHz to
allow clients to connect on 240 MHz.

Change-Id: I22ebfe2f266ee9c69fd2991779dde8307fb71f4c
Signed-off-by: Aloka Dixit <aloka.dixit@oss.qualcomm.com>