UPSTREAM:CP wifi:ath12k Decrement ahsta->num_peer during link sta unassignment
There is a gap in the logical_link_idx sent in the peer_assoc command.
For a two-link association, it should be 0 and 1. However, due to a
failure in creating link station (which was assigned link_idx 1) caused
by an AST failure, the next link creation is assigned link_idx 2.
This count should be decremented whenever a link station creation fails.

As a result, potential incorrect IPC messages are sent across the chipset
on FW, leading to unexpected behavior.

Decrement ahsta->num_peer which is assigned as link station logical_link_idx
during ath12k_mac_unassign_link_sta().

Patch-dependency: None
Patch-work: None

Change-Id: I16aef805c47282fbb83834392582f77a4ce37c67
Signed-off-by: Vignesh C <quic_vignc@quicinc.com>