UPSTREAM:CP wifi:ath12k: Fix the unassign vif during the deauth
Here repeater is receiving the deauth from the hostapd since the root AP is
going down. In that instance repeater will go for simulated assert using CLI
command which will avoid the unassign vif for the asserted SoC. So, asserted
SoC will have stale entry of arvif which needs to be removed during the
deauth case.

Fix is to stop the vdev and remove the interface of the asserted SoC before
going for recovery. This will help in deleting all stale entry of the arvifs
before going for recovery.

Patch-dependency: 811-wifi-ath12k-change-to-stop-peer-delete-the-primary-l.patch
		  767-ath12k-Check-if-ar-is-down-before-handling-mac-opera.patch
Patch-work: None

Change-Id: I9054dac3b963a0022e7025680d6120ab0b230d09
Signed-off-by: Gautham Kumar Senthilkumaran <quic_gauthamk@quicinc.com>
(cherry picked from commit 0c22e9964457246eb306162eeadbefcf9816d678)