QSDK:PROTOCOL: fix dead lock in color change bss notify
In the current mac80211 design, when a BSS color change request is
received for a tx BSS, it updates all non-tx BSSes in the same MBSSID
group.

A deadlock occurs when the BSS color change finalize worker is
scheduled for the TX BSS, and at the same time, a stop AP command is
received for a non-TX BSS in the same MBSSID group.

Let's assume wlan0 is the TX BSS and wlan1 is the non-TX BSS in the same
MBSSID group. When the color change finalize worker is scheduled for wlan0,
it acquires the wlan0 sdata lock and the local mtx lock (radio-specific),
and notifies the driver of the color change.
While looping to update the color change for the non-TX BSS,
if a stop AP command is received for wlan1, it acquires the wlan1 sdata
lock and waits for the local mtx lock, which is already acquired by
the color change finalize worker. When updating the color change for
wlan1, it waits for the wlan1 sdata lock, which is already taken by
the stop AP command, resulting in a deadlock.

Avoid deadlock by implementing offload logic to update the color change
for non-TX BSS whenever a color change request is received for TX BSS.

Change-Id: If36c509cd560d774016f9491be62430af84b1984
Signed-off-by: Ramasamy Kaliappan <quic_rkaliapp@quicinc.com>