UPSTREAM:CP wifi:ath12k Send peer_delete for self peer before vdev delete
It has been observed that the vdev delete command is sent to the
firmware before the self peer delete command. In cases where the
scan vdev delete work is queued before it gets scheduled, the
assign_vif_chanctx() is received to create the link. In this scenario,
the HOST will attempt to force clean the scan vdev and create a
new vdev with a new chanctx(). Here, the vdev_delete is sent without
the self peer delete.

Ensure that the peer delete for the self peer is sent in AP mode
before sending the vdev delete to the firmware

Patch-dependency:
	896-wifi-ath12k-STA-mode-multiple-scan-support.patch

Change-Id: I5566d32c1fe90ed48317653bebb3d39e0c1f2d8c
Signed-off-by: Vignesh C <quic_vignc@quicinc.com>