UPSTREAM:PROTOCOL wifi: mac80211: Cancel CSA/CCA in ML Reconfig work in STA
Currently, when a channel switch with block tx set and
ML reconfigure link removal is triggered on same link with
less tbtt which results in the link to be removed first.
This scenario leads to tx to be blocked - ping fails.
This happens because block_tx is set during
CSA, and the vif_queues are stopped during the channel switch.

Since the link is removed before the channel switch completes,
block_tx should be unset before removing the link.

To fix this issue, wake up the queues and allow them to continue
transmitting data before removing the link.
Also, unset the CSA and CCA flags before removing the link.

Patch-dependency: ML Reconfiguration Link Removal

Patch-work: None

Change-Id: I7930a9b5ab3e36469cae91b9f6995eb46a4b06f9
Signed-off-by: Yuvarani V <quic_yuvarani@quicinc.com>