QSDK:CP:mac80211.sh Update frequency calculation before calculating center freq
Currently, when a new link is dynamically added through 'mlo_add_link'
option when there is no BSS on that radio created before, script will
call the network reload and proceed for creating the hostapd config
through 'drv_mac80211_setup' callback. In this flow, frequency value is
not updated properly, hence the seg0 index calculation is failing which
results in wrong center freq getting selected.

This is due to current implementation, where the channel to frequency
conversion is happening only when the setup is requested for normal
network restart and not for dynamic link addition.

Fix this by updating the frequency irrespective the setup flow
requested.

Fixes: Ifaa8e76bc903 ("Add support to handle mlo links addition")

Change-Id: I7450d1a7c9ba92807208deb053483139d0c08155
Signed-off-by: Manish Dharanenthiran <quic_mdharane@quicinc.com>