QSDK: wifi: ath12k: Add callbacks in arch_ops for rx APIs
Add callbacks in arch_ops for rx APIs which are defined
in wifi7 and need to be called from the common code.
This is required, as these APIs will be part of
ath12k_wifi7.ko and the common module ath12k.ko needs
ops mechanism to call these APIs.


Signed-off-by: Harsh Kumar Bijlani <quic_hbijlani@quicinc.com>
Change-Id: I07be2c1e3f61663baba7525025e1aa1d77d3a569