wifi: ath12k: HAL changes for Unified Driver Framework
The primary objective of this set is to modify the HAL layer
to suit the multi-KO model. The HAL APIs will be implemented
differently as required by different chipsets and these
implementations will be abstracted using an ops mechanism
(by registering function pointer callbacks). The hw specific
implementations will be registered as callbacks in the hal_ops.

			common hal API
				|
				|
				|
				|
	=============================================
			     hal_ops
	=============================================
	 _______________              _______________
	|       |       |            |       |       |
	|       |       |            |       |       |
	| qcn   |  wcn  |            | qcn   |  wcn  |
	| chip  | chip  |            | chip  | chip  |
	|_______|_______|            |_______|_______|

	  ath12k_wifi7.ko             ath12k_wifi8.ko

Also, the hal structures that are specific to HW (wifi7) are
kept within hal headers in wifi7 directory and the structures
that are common are placed in the headers present in the ath12k
directory.

Change-Id: If48b333036d327ecad1608a601eafe8096ae8469
Signed-off-by: Pavankumar Nandeshwar <quic_pnandesh@quicinc.com>
Signed-off-by: Ripan Deuri <quic_rdeuri@quicinc.com>