QSDK: mac80211: Fix warn on if mesh metric is zero
During performence measurement, stats is disabled through boost
performence script. So, host is not receiving the metric values
from firmware and the value becomes zero during mpath finding.
This issue addressed through static mpath. But,in RvR far range
test, connection becomes unstable and getting this warnon
continuously. So, add a check to prevent this warnon if the
stats is disabled.

Patch-dependency:
	0206-wifi-mac80211-avoid-tx-pending-queue-in-native-wifi-.patch

Change-Id: Iba0e74614f9b9b531e394db02e1801071d0555e0
Signed-off-by: Thiraviyam Mariyappan <tmariyap@qti.qualcomm.com>
Signed-off-by: Ramanathan Choodamani <rchoodam@qti.qualcomm.com>