QSDK:PLATFORM: wifi: ath12k: Enable amss_dualmac.bin loading for QCN9274
Add changes to load amss_dualmac.bin when firmware-2.bin is not present.

If firmware-2.bin fails to load, fw features will not be populated, so
any checks on fw features should be skipped in fallback scenarios. So
adding a check on fw api version to bypass any fw features related checks.

Change-Id: I26df32c667a9a851aa0d60ad883169d7c6532125
Signed-off-by: Avula Sri Charan <asrichar@qti.qualcomm.com>