UPSTREAM:wifi:DP:ath12k: fix for wds multicasat and vif tcl metadata
During AP-STA repeater scenario, initially,
AP interface is brought up with vdev_id 0 and
in parallel, STA interface is brought up with
vdev_id 1 and scan is initiated for STA interface.
While the scan is on going, AP interface is disabled
as the STA is not yet associated. Hence, vdev_id 0 is
deleted.

Once STA scan is completed, vdev (vdev_id 1) is deleted
and during association again STA interface is created with
vdev_id 0. During this, the vdev_id information is appended
in tcl_metadata which results in incorrect configuration of
vdev_id because the ahvif's deflink still had the old vdev_id
in the tcl_metadata information.

In host we set WMI_VDEV_PARAM_AP_ENABLE_NAWDS when we configure a
interface as WDS AP and WDS sta connecting to the AP in eth mode.
Due to this flag fw expect all the multicsat traffic should come as
4addr frame, if the AP interface sends any multicast frame as
3addr frame type fw return the tx completion status as
HAL_WBM_REL_HTT_TX_COMP_STATUS_REINJ and leads to multicast
traffic failure for the 3addr clients connected to the AP.
To fix this by setting host inspected meta data flag for the
3addr multicast frame in driver before queueing the packet to
firmware/hardware will avoid fw returning to host with reinject
status and allow multicast traffic.

Original patches: 717-ath12k-fix-tcl_metadata-vdev_id-info.patch
971-wifi-ath12k-Fix-multicast-traffic-issue-on-WDS-mode-AP-in.patch

Change-Id: Ib477cebd0e2832a0efc414f57663774052a5e668
Signed-off-by: Tamizh Chelvam Raja <tamizhr@qti.qualcomm.com>