QSDK:PLATFORM: wifi: ath12k: Free DP irq in deinit path
All interrupts in ahb are registered usning devm_request_irqa for
which kernel will take care to free on driver_unregister. Because of
which freeing irq for ahb was never taken care by ath12k drver for
ahb.

With new design the DP interrupts will be re-registered on Q6 restart and in this
case driver have to take care of cleaning those interrupts before Q6 restart.

Change-Id: Ic66167757cf4a01667a0a02be2229a8e438086fa
Signed-off-by: Avula Sri Charan <asrichar@qti.qualcomm.com>