QSDK:CP: wifi: mac80211: Backport upstream AP_VLAN interface link handling
Design for AP_VLAN interface links in upstream is different than
internal code. Backporting upstream changes that implement links for
AP_VLAN interfaces.

First, remove existing link handling for AP_VLAN interface before
adding upstream changes.

Bring in upstream design of AP_VLAN links from [1]. Part of the
upstream changes were brought in as part of split-phy change but not
all.

Add checks for use_4addr flag when checking SUPPORTS_VLAN_DATA_OFFLOAD
flag in mac80211. This is a WAR for current issue that the
SUPPORTS_VLAN_DATA_OFFLOAD hw flag is utilized for both Dynamic VLAN
and WDS and there is no way to make a distinction between the two just
based on the hw flag. This is only a temporary internal solution
because currently VLAN offload is only supported for WDS in ath12k,
but this might not be the case in upstream. It is possible that
Dynamic VLAN offload might be supported instead.

With new design brought to internal code, MLO WDS is not functional.
Add additionaly functionality to AP_VLAN interface's handling of links
for MLO WDS.

[1] Link: https://lore.kernel.org/linux-wireless/20250325213125.1509362-1-muna.sinada@oss.qualcomm.com/

Patch-dependency:

0007-wifi-mac80211-Create-separate-links-for-VLAN-interfa.patch
0141-wifi-mac80211-Changes-for-WDS-MLD.patch
0207-WAR-wifi-mac80211-Assign-AP_VLAN-link-to-null-when-ROOT-AP-in.patch
0337-01-QSDK-DP-wifi-Add-support-for-PPE-DS-Direct-switch-fe.patch

Change-Id: Iac38b060ca04b99e5f62f94b05d702b34bec1d9a
Signed-off-by: Muna Sinada <muna.sinada@oss.qualcomm.com>