QSDK:ath12k:DP: Fix memory leak in tx completion handler
The current code is not adding the used descriptor to the free
list when receiving the descriptor from the firmware, making the
descriptor unusable for the next iteration and causing a memory
leak. Hence, fixing this issue by moving the descriptor to the
descriptor free list.

Additionally, removed the rcu readlock for tx desc buffer release
which is not required.

Change-Id: Id9dbe507a6d294a8d3144ec42d1d4c0abb6d4ed9
Signed-off-by: Thiraviyam Mariyappan <tmariyap@qti.qualcomm.com>