UPSTREAM: DP: wifi: ath12k: Fix Issues in REO RX Queue Updates
Currently, under stress scenarios, when the REO command ring is
full, the peer delete update for the RX queue fails due to lack
of space to send the command. In such cases, the host performs
a dma_unmap and frees the allocated memory. However, the hardware
still retains a reference to the same virtual address. During the
next allocation cycle, the kernel may reassign this address,
unaware that the hardware is still using it. This can lead to
memory corruption, as the host may access or write to memory
that is still in use by the hardware.

To prevent memory corruption, a retry mechanism is added for
HAL_REO_CMD_UPDATE_RX_QUEUE. This involves introducing a new
list in the dp structure, protected by a dedicated lock to
ensure safe access. With this change, the host no longer frees
memory immediately on failure. Instead, the memory is freed
only after the hardware has released the corresponding virtual
address.

Additionally, the DP_REO_CMD_RING_SIZE is increased to 256 to
reduce the likelihood of ring exhaustion. Checks are also added
for rx_tid->vaddr during reo_cmd_list_cleanup() and
dp_reo_cmd_free() to prevent double free or unmap operations.
Furthermore, a 1KB cache flush command is now used for QoS TID
descriptors to improve efficiency.

Below patches changes are ported onto 4.0 :
649-ath12k-fix-double-free-of-peer-rx_tid-during-reo-cmd.patch
703-ath12k-Update-pending-desc-size-in-reo-cache-flush.patch
672-ath12k-Fix-freeing-hal-reo-status-entry-in-host.patch
750-ath12k-Add-retry-mechanism-for-update_rx_queue-reo-cmd.patch

Currently, these patches are in internal review for upstream
with title "Fix Issues in REO RX Queue Updates".

Change-Id: I12c8658df1d06ea99f387a9aa00a42bf696e2ebd
Signed-off-by: Manish Dharanenthiran <mdharane@qti.qualcomm.com>
Signed-off-by: Nithyanantham Paramasivam <nithp@qti.qualcomm.com>