UPSTREAM: CP: wifi: ath12: Copy eid into skb control buffer
While copying the SKB into the ring, driver has to copy the end point
into skb control buffer. Host depends on this eid to remove the SKB from
ring and wake up the tx queue. This is crucial for next WMI cmd to be
copied to CE ring.

Currently the eid is not copied while enqueuing the SKB which caused eid
mismatch upon tx completion interrupt.

Change-Id: Ib2f26edcd0feb6d9546e3461fb7880faf1d1c52f
Signed-off-by: Ramya Gnanasekar <rgnanase@qti.qualcomm.com>