UPSTREAM:DP: wifi: ath12k: Add DP UMAC Reset support
It contains the base changes to add support for umac dp
reset of partner SOC when one SOC crashes in MLO
as a part of MODE 1 SSR

Enable umac reset for IPQ5332, QCN6432 and IPQ5424

Add support to parse the dts for umac_reset interrupt line
and enabling it for IPQ5332, QCN6432 and IPQ5424.

In current implementation entire ring mask is being set to 0 during
umac reset to avoid processing dp interrupts, instead of that
disabling the dp interrupts during pre_reset_start state and then
enabling them back after post_reset_complete of umac reset.

As the umac reset will take place in atomic context bypassing
napi_sync by checking on the flag ATH12K_FLAG_UMAC_PRERESET_START
to avoid sleep from atmoic context issue.

QCN6432 support:
As the Q6 of ipq5332 is shared between all the radios of qcn6432
the umac reset interrupt line needs to be resetted for qcn6432
every time after receiving the interrupt so that the next interrupt
can be raised for that reason writing the value 0x00000100 into a pci
slot specific address.

Currently for qcn6432 MSI based interrupts are being used but as per the
design one interrupt line is reserved for umac_reset interrupt so adding
support to parse it from the dts.

In this process shared memory will be used by Host and
Firmware to communicate the sequence of events. Host will
clear the data structures such as tx/rx descriptors and
re-initialize the rings and reset hp/tp to 0.

Following patches are partially implemented as part of this change:
(Partially implemented as few of the patches are already
added as part of package upgrade)

800-wifi-ath12k-Add-UMAC-DP-reset-support.patch
850-03-wifi-ath12k-avoid-pass-the-number-of-entry-to-replen.patch
(partial)
850-06-wifi-ath12k-avoid-used-list.patch (partial)
900-wifi-ath12k-Enable-umac-reset-for-IPQ5332.patch
901-wifi-ath12k-Enable-umac-reset-for-QCN6432.patch
955-02-ath12k-simplify-rx-buffer-ring-replenish-logic.patch (Partial)
960-002-wifi-ath12k-Add-hw-params-configuration-for-ipq5424.patch(umac reset specific changes)

Change-Id: Ic1f4d2d0311b955060b25498288e31ac0854a9be
Signed-off-by: Adwait Nayak <adwaitn@qti.qualcomm.com>