UPSTREAM:PROTOCOL:wifi:ath11k:Add support for Tx/Rx chainmask based on radio id
Add support to add the radio id parameter for which Tx/Rx chainmask
needs to be changed from userspace. Pass on this radio index to
underlying drivers as an additional argument.

A value of NL80211_WIPHY_RADIO_ID_MAX(255) indicates radio index is not
mentioned and the configuration applies to all radio(s) of the wiphy.

Change-Id: I0e36d84605a9989688d6ccee24894507261c7770
Patch-work: None
Patch-dependency:patches/subsys/0287-UPSTREAM-PROTOCOL-wifi-mac80211-cfg80211-Set-get-wip.patch
Signed-off-by: Uraj Sasan <usasan@qti.qualcomm.com>