UPSTREAM: DP: wifi: ath12k: Monitor HAL changes for Unified Driver Framework
The primary objective of this set is to modify the monitor HAL layer
to suit the multi-KO model. The monitor HAL APIs will be implemented
differently as required by different chipsets and these implementations
will be abstracted using an ops mechanism
(by registering function pointer callbacks). The hw specific implementations
will be registered as callbacks in the hal_mon_ops

Change-Id: I8d73c09e29271a700f64e5c4c3ca6cdf0fd27490
Signed-off-by: Alok Singh <aloksing@qti.qualcomm.com>