UPSTREAM:DP: wifi: ath12k: Add vendor event for primary link migration notification
In scenarios where the user application is responsible for load
balancing and communicating the destination MAC address (MLD peer MAC
address) and its primary link ID to the Ethernet-side QCA driver, there
is a need for a notification mechanism to ensure the primary link
changes in WLAN subsystem are communicated to user applications and also
to manage the load of that primary link in a better way.

Add vendor NL event to address above requirement by providing a way
for the user application to be notified about primary link migrations,
allowing it to update its configuration and ensure that other
subsystems, such as Ethernet, use the correct link for communication.

Add code changes to extend the usage of the vendor command
QCA_NL80211_VENDOR_SUBCMD_PRI_LINK_MIGRATE as an event to notify user
applications when the primary link has changed.

This event will send the MLD MAC address of the peer and the new link ID
which is set as the primary link.

Also modify the existing approach to parse and extract the attributes.
Use NL attributes and fetch the required vendor command data.

Patch Dependency: UMAC migration related patch
Feature: UMAC migration

Change-Id: I143479c0041df0c1a1220738b0faed18cbbbc62f
Signed-off-by: Logapriya P <logapriy@qti.qualcomm.com>