UPSTREAM:PROTOCOL: wifi:ath12k: fix firmware crash when removing ML station with active link
Current code skips peer cleanup in ath12k_mac_station_remove() for ML
(Multi-Link) stations. This has a problem when a link is removed while
the station is still associated, as the AP sends a vdev delete command
without cleaning up the peer, leading to a firmware crash due to
non-zero peer count.

We can improve this by allowing peer cleanup for ML stations during
ML reconfiguration, because it ensures the peer count is properly
decremented and prevents firmware crashes.

Change-Id: I87cd7a78f1b7f725825868b0958a7d2ed19fbe00
Signed-off-by: Karthik M <karm@qti.qualcomm.com>