UPSTREAM: wifi: ath12k: Update vdev params to send duplicate beacon
In 6GHz to make scan faster for the clients, duplicate beacon will be
advertised in PSC channels when AP is operating in non-PSC channel.
Firmware expects a VDEV param WMI_VDEV_PARAM_6GHZ_PARAMS for 6GHz in
this case to advertise duplicate beacon in 20MHz.

But the vdev param is skipped for transmitting vdev due to usage of un
initialized boolean variale.
Fix the issue by initializing the boolean variable used for validation

Fixes: Ie484739288 (CP: wifi: ath12k: PKG upgrade changes set 3)

Change-Id: I2dfc4e786f29d373d2e6f564a3fe1282f58f90bd
Signed-off-by: Ramya Gnanasekar <rgnanase@qti.qualcomm.com>