mac80211: Fix for use-after-free in tx path
key memory is allocated and stored in two places,
  1. sta->ptk
  2. sta->fast_tx

After freeing the key memory, NULL pointer is assigned to sta->ptk but
not assigning to sta->fast_tx. Due to this, freed key memory accessed
in the tx path. To fix this issue, set tx key for old key after freed
the key memroy.

Change-Id: Id07a866226d74cbd8099e5912fd46d61893eebeb
Signed-off-by: Balamurugan Ramar <bramar@qti.qualcomm.com>