UPSTREAM: PROTOCOL: wifi: cfg80211/mac80211: Add support for advertised TTLM and TTLM event notifications
An AP MLD may advertise a mandatory TTLM by including a TID-To-Link Mapping
element in the Beacon and Probe Response frames that the APs affiliated
with the AP MLD transmit. An AP that advertises a TTLM shall include the
Mapping Switch Time field and shall set it to the time, in units of TUs,
of the TBTT of a DTIM beacon of one of the APs affiliated with the AP MLD.

Beginning at the time indicated in the Mapping Switch Time field, the
indicated TTLM is established and the Mapping Switch Time field is no
longer included.

An AP that advertises a TTLM shall map all TIDs to the same link set in
the advertised TTLM, both for DL and UL. The Direction field of an
advertised TID-To-Link Mapping element shall be set to 2.

An AP MLD shall include two TID-To-Link Mapping elements in the Beacon and
Probe Response frames that the APs affiliated with the AP MLD transmit, if
there is already an established advertised TTLM and the AP MLD intends to
replace it with a nondefault advertised TTLM. In this case, the AP MLD
shall not include the Mapping Switch Time field in the currently
established advertised TID-To-Link Mapping element, and shall include the
Mapping Switch Time field in the TID-To-Link Mapping element, in order to
indicate an advertised TTLM that will be established in the future.
The value of the Expected Duration field of the existing TID-To-Link
Mapping element shall indicate a remaining duration that ends at the same
time as indicated by the Mapping Switch Time field of the TID-To-Link
Mapping element.

This patch series introduces enhancements to the cfg80211 and mac80211
subsystems to support the advertised TTLM on AP interface type in driver
offloaded mode.
With driver offloaded mode of advertisement of TTLM,
  - TTLM element inclusion in beacon is offloaded.
  - TTLM element inclusion in Probe Response, (Re)Assoc Response is
not offloaded. Element is formed and included in userspace application,
however, the Mapping Switch Time(in TSF) before advertisement starts is to
be read from driver. And the current ED to include in probe response frame
is to be read from driver. This MST and ED dependency on driver makes the
TTLM advertised in Probe/(Re)Assoc response partially offloaded.
  - Mapping Switch Time conversion from TU to TSF is offloaded. The
converted TSF value is notified to upper layers for further inclusion in
probe response frames.
  - Expected Duration countdown for every beacon is offloaded.
  - Upper layers are notified when Mapping Switch Time and Expected
duration timers are expired.

Patch-dependency: 0306-0001-UPSTREAM-PROTOCOL-wifi-nl80211-Extend-the-set_ttlm-to-support.patch

Feature: TTLM

Change-Id: I834b5d174dbc9e6b8f210effc78cc47476033b1e
Signed-off-by: Amith A <amitajit@qti.qualcomm.com>
Signed-off-by: Shwetha G K <kshwetha@qti.qualcomm.com>