UPSTREAM:PROTOCOL: wifi: ath12k: Add driver offload advertised TTLM support
This patch series adds the support in ath12k driver to start the advertised
TTLM on AP interface and also adds the support to handle the TTLM status
indication from firmware via WMI event.

An AP MLD may advertise a mandatory TTLM by including a TID-To-Link Mapping
element in the Beacon and Probe Response frames that the APs affiliated
with the AP MLD transmit. An AP that advertises a TTLM shall include the
Mapping Switch Time field and shall set it to the time, in units of TUs,
of the TBTT of a DTIM beacon of one of the APs affiliated with the AP MLD.

After starting the offloade advertised TTLM on AP MLD, firmware updates few
status to driver via WMI event indicating
  - MST TSF update - contains MST in TSF to be added in probe response's
 TTLM element if any is going to be sent from application.
 - MST expiry: Indicates that the mapping is effective as MST expired.
 - ED expiry: Indicates that already established mapping to which MST
expiry was notified earlier is no longer effect, and if any upcoming TTLM
request is queued, then that mapping will be effective.

Patch-dependency:
0381-0001-UPSTREAM-PROTOCOL-wifi-ath12k-Add-WMI-framework-to-s.patch
0381-0003-UPSTREAM-PROTOCOL-wifi-ath12k-Add-TTLM-offload.patch
0415-0002-UPSTREAM-CP-wifi-ath12k-Enable-QoS-support-in-the-Fi.patch

Feature: TTLM

Change-Id: Ie0ebaf83ebd38925495c759e188c58a874becfc8
Signed-off-by: Amith A <amitajit@qti.qualcomm.com>