UPSTREAM:PROTCOL: wifi: cfg80211/mac80211: Add supported power mode and EIRP power info to NL80211_CMD_GET_WIPHY
Extend the response of the NL80211_CMD_GET_WIPHY command to include
supported power modes and EIRP values for every 6 GHz channel.
If AFC is completed, the EIRP value to be programmed into the hardware
for SP power mode should be the minimum of the EIRP value obtained in
the AFC response and the regulatory EIRP value. Additionally, add a
mac80211_config_ops, 'get_afc_eirp_pwr', to fetch the EIRP value received
in the AFC payload for a given frequency.

Fixes: None
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Path-dependency:
ath12k/0346-02-wifi-ath12k-Add-support-for-processing-AFC-power-eve.patch
ath12k/0346-06-wifi-ath12k-Extract-AFC-response-and-send-AFC-cmd-to.patch

Change-Id: Ib642b6d495d51dc6d1f5b6d209c2d2e556ffba1f
Signed-off-by: Amith A <amitajit@qti.qualcomm.com>