UPSTREAM:PROTOCOL: wifi: cfg80211/mac80211: Enhance TTLM negotiation support
The TID-to-Link Mapping (TTLM) mechanism allows an AP MLD or non-AP
MLD that has performed or is performing multi-link setup to determine
how uplink (UL) and downlink (DL) QoS traffic corresponding to TID
values between 0 and 7 will be assigned to the setup links for the
non-AP MLD. By default, all TIDs are mapped to all setup links for
both UL and DL.

An AP MLD and a non-AP MLD can negotiate with each other to assign
links to each TID based on end-user requirements. The mapping can be
either homogeneous or heterogeneous. In homogeneous mapping, all TIDs
are mapped to the same set of links. In contrast, in heterogeneous
mapping, each TID can be mapped to a different set of links.

TTLM negotiation capability is advertised as part of the basic
multi-link element by both the AP MLD and the non-AP MLD.

The negotiation is carried out using TTLM action frames, as specified
in P802.11be_D3.0 Section 35.3.7.1.3 Negotiation of TID-to-link mapping

When TTLM negotiation is performed, a TID can be mapped to a subset
of the established links, ranging from a single link to all the links
that have been set up.

A setup link is considered enabled for a non-AP MLD if at least one
TID is mapped to that link in either the DL or UL direction. If no
TIDs are mapped to a link in either direction, that link is
considered disabled. At any point in time, a TID must always be
mapped to at least one setup link for both UL and DL. By default, all
setup links are enabled.

This patch series introduces some enhancements made in existing TTLM
negotiation within mac80211 and cfg80211 subsystems.

Key changes include:
Enhancements to TTLM negotiation to support AP
interface. Current infrastructure allows the SET_TTLM to be triggered
on STA interface alone. Extend the framework to allow AP interface as
well where the actual negotiation action frames generated externally
in AP mode, and this infra lets driver handle the actual TIDs mapping.
Incase of negotiation on AP, current active links / vif cfg not changed
as the negotiation vary per peer.

Patch-dependency: None

Feature: TTLM

Change-Id: Id4a16cb22d0b55b658eda32475d9b4d62bf7bc6f
Signed-off-by: Shwetha G K <kshwetha@qti.qualcomm.com>
Signed-off-by: Dhanavandhana Kannan <dhanavan@qti.qualcomm.com>