UPSTREAM: PROTOCOL: wifi: ath12k: Fix AFC request for split radio
Currently, when forming the AFC request frequency range, the number
of frequency range is set to the number of regulatory SP frequency
ranges supported for the current country.
For US, this number is 2.

But, the actual frequency range is constructed by the intersection
of the regulatory SP range abd the hardware frequency range
supported by the radio. Thus, for a 6 GHz High or Low band radio,
the actual frequency range is 1, as the hardware supports only
one of the two regulatory SP ranges for the US.

Thus, we end up sending an AFC request with number of frequency
ranges set to 2, but with only one frequency range actually
provided. This casuses the AFC server to reject the request
with an error.

To fix this, set the number of frequency ranges to based on the
number of actual frequency ranges constructed, instead of
the number of regulatory SP ranges supported for the country.

Fixes: fb0c797a1809 ("UPSTREAM: PROTOCOL: Add AFC Expiry event changes")
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
0373-d-UPSTREAM-PROTOCOL-Add-APIs-to-construct-AFC-frequenc.patch

Change-Id: I2766c9f1708a211e37a1c4423afcb18b70f7f7ed
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>