QSDK:CP:wifi:cfg80211: Fix for SW progressive DFS
While checking for start and end frequency of device, bandwidth was in
mhz and center frequency was in KHZ causing miscalculation. Fix it by
converting bandwidth to KHZ.

When radar was detected when software progressive DFS is enabled, only
radar infected channel should be marked unusable. In current code, all
the subchannels are marked NOL and hence channel switch fails.
Adding a WAR to avoid setting all subchannels are unusable.

Change-Id: I60754b67831c3567b21e2a9d324a255fa9700d33
Signed-off-by: Paulson Raja L <lraja@qti.qualcomm.com>
Signed-off-by: Ramya Gnanasekar <rgnanase@qti.qualcomm.com>