QSDK:CP:wifi:ath12k: Fix double free of skb
In ath12k_wmi_force_fw_hang_cmd() because of an un-initialized ret var
the skb is getting freed in error path though it is successful. Later this
same skb will get freed in ath12k_htc_control_tx_complete() which leads
to double free of skb and causing issues in skb alloc/free path randomly.

Fix this by initializing the ret to 0.

mac_op_change_vif_links() is expected to be called while reconfig is in
progress and in that time arvif is expected to be present so skip the
WARN_ON during recovery.

Also clear the ATH12k_FLAG_RECOVERY after the mac_op_reconfig() done
instead of doing it immediately after core_restart().

Use ahsta->ml_peer_id instead of dp_peer->peer_id in dp_peer_cleanup().

TODO: In upstearm we don't see this Warning's during recovery so need to
check on the difference and see if we can remove the check's on FLAG_RECOVERY
in mac_op_change_vif_links();

Change-Id: Ia3217d85c0409ce817e7d19927b233c717be9e43
Signed-off-by: Avula Sri Charan <asrichar@qti.qualcomm.com>