UPSTREAM:PROTOCOL:wifi:ath12k: Fix Invalid payload for dual 6 GHz AFC
Currently, when we receive an invalid or empty AFC payload,
we mark the SP frequency ranges supported by the current radio's
hardware range as NO IR in the regd structure for the current
radio.

But, this does not mark those frequency ranges as NO IR
in the regd structure of the partner 6 GHz radio in the same
wiphy. Thus, when regd update happens through the partner
radio, the SP frequency ranges for the radio which received
the invalid AFC payload are not marked as NO IR.

To fix this, update the regd structure of the partner radio
as well when we receive an invalid AFC payload.

As an optimization, do not update the regd structure with
the NO IR frequencies if the regd structure already has
those frequencies marked as NO IR.

Add debug messages to log the regd updation due to AFC
payload processing.

Fixes: 86ecd262bcc8 ("UPSTREAM:PROTOCOL:wifi:ath12k: Add 6GHz SP NO_IR Support")
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
subsys/0048-cfg80211-mac80211-regulatory-awgn-and-6g-channel-cha.patch
ath12k/0346-02-wifi-ath12k-Add-support-for-processing-AFC-power-eve.patch
ath12k/0346-06-wifi-ath12k-Extract-AFC-response-and-send-AFC-cmd-to.patch

Change-Id: I9ed37258430f501fa8e3166883f6f1d40216c687
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>