UPSTREAM:PROTOCOL:wifi: ath12k: Fix TTLM tlv header addition for legacy cases
Currently memory is allocated for TTLM header, but when
legacy client tries to associate, TTLM header is not added
resulting in length mismatch of WMI message in firmware.

Fix this by including the TTLM header for legacy cases as well.

Patch-Dependency: 0381-0004-UPSTREAM-PROTOCOL-wifi-ath12k-WMI-
                  changes-to-send-ttlm-params-in-peer.patch

Patch-work: None

Fixes: I68dc41c7522(UPSTREAM:PROTOCOL: ath12k: Add TTLM negotiation
and WMI support for peer mapping)

Change-Id: Ic1134e65b60e9410b6ef388f67e763b50f50bd24
Signed-off-by: Shwetha G K <kshwetha@qti.qualcomm.com>
Signed-off-by: Dhanavandhana Kannan <dhanavan@qti.qualcomm.com>